+++ backend/channel_plugin/apps/channelmessages/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ChannelmessagesConfig(AppConfig):
+    name = "channelmessages"
