+++ backend/channel_plugin/apps/channels/serializers.py
@@ -1,7 +1,10 @@
+from apps.roles.serializers import RoleSerializer
 from django.utils.text import slugify
 from rest_framework import serializers
 
-from .models import Channel, ChannelMessage, Role, Thread
+from channel_plugin.utils.customrequest import Request
+
+from .models import Channel
 
 
 class ChannelSerializer(serializers.Serializer):
@@ -14,6 +17,15 @@ class ChannelSerializer(serializers.Serializer):
         """
         Validate name doesnt alreat exist in organization
         """
+        data = {"name": name}
+        response = Request.get(self.context.get("org_id"), "channel", data)
+
+        try:
+            if response.json():
+                raise serializers.ValidationError({"error": "Name already exist"})
+        except AttributeError:
+            if response:
+                raise serializers.ValidationError({"error": "Name already exist"})
         return name
 
     def to_representation(self, instance):
@@ -21,30 +33,6 @@ class ChannelSerializer(serializers.Serializer):
         slug = slugify(instance.get("name"))
         channel = Channel(**instance, slug=slug)
         data = {"channel": channel}
-        return data
-
-
-class Permission(serializers.Serializer):
-
-    name = serializers.CharField(max_length=50)
-    description = serializers.CharField()
-
-
-class RoleSerializer(serializers.Serializer):
-
-    name = serializers.CharField(max_length=10)
-    channel_id = serializers.CharField(read_only=True)
-    permissions = Permission(many=True)
-
-    def to_representation(self, instance):
-        instance = dict(instance)
-        channel_id = self.context.get("channel_id")
-        instance["permissions"] = [
-            {**item, "key": slugify(item.get("name"))}
-            for item in instance.get("permissions")
-        ]
-        role = Role(**instance, channel_id=channel_id)
-        data = {"role": role}
         return data
 
 
@@ -66,76 +54,24 @@ class ChannelUpdateSerializer(serializers.Serializer):
         """
         Validate name doesnt already exist here
         """
+        data = {"name": name}
+        response = Request.get(self.context.get("org_id"), "channel", data)
+        try:
+            if response.json():
+                raise serializers.ValidationError({"error": "Name already exist"})
+        except AttributeError:
+            if response:
+                raise serializers.ValidationError({"error": "Name already exist"})
         return name
 
     def to_representation(self, instance):
         if instance:
             instance = dict(instance)
-            channel = Channel(**instance)
+            slug = slugify(instance.get("name"))
+            channel = Channel(**instance, slug=slug)
             data = {"channel": channel}
             return data
         return super().to_representation(instance)
-
-
-class ChannelMessageSerializer(serializers.Serializer):
-
-    user_id = serializers.CharField(max_length=10, required=True)
-    content = serializers.CharField()
-    timestamp = serializers.DateTimeField(read_only=True)
-
-    def to_representation(self, instance):
-        instance = dict(instance)
-        channel_id = self.context.get("channel_id")
-        message = ChannelMessage(**instance, channel_id=channel_id)
-        data = {"channelmessage": message}
-        return data
-
-
-class ChannelMessageUpdateSerializer(serializers.Serializer):
-
-    user_id = serializers.CharField(read_only=True)
-    channel_id = serializers.CharField(read_only=True)
-    content = serializers.CharField()
-    emojis = serializers.ListField(serializers.CharField(), allow_empty=True)
-    pinned = serializers.BooleanField(default=False)
-    edited = serializers.BooleanField(default=False)
-    timestamp = serializers.DateTimeField(read_only=True)
-
-    def to_representation(self, instance):
-        instance = dict(instance)
-        message = ChannelMessage(**instance)
-        data = {"message": message}
-        return data
-
-
-class ThreadSerializer(serializers.Serializer):
-
-    user_id = serializers.CharField(max_length=10, required=True)
-    content = serializers.CharField()
-    timestamp = serializers.DateTimeField(read_only=True)
-
-    def to_representation(self, instance):
-        instance = dict(instance)
-        channelmessage_id = self.context.get("channelmessage_id")
-        thread = Thread(**instance, channelmessage_id=channelmessage_id)
-        data = {"thread": thread}
-        return data
-
-
-class ThreadUpdateSerializer(serializers.Serializer):
-
-    user_id = serializers.CharField(read_only=True)
-    channelmessage_id = serializers.CharField(read_only=True)
-    content = serializers.CharField()
-    emojis = serializers.ListField(serializers.CharField(), allow_empty=True)
-    edited = serializers.BooleanField(default=False)
-    timestamp = serializers.DateTimeField(read_only=True)
-
-    def to_representation(self, instance):
-        instance = dict(instance)
-        thread = Thread(**instance)
-        data = {"thread": thread}
-        return data
 
 
 class SearchMessageQuerySerializer(serializers.Serializer):
