+++ backend/channel_plugin/apps/roles/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class RolesConfig(AppConfig):
+    name = "roles"
