+++ backend/channel_plugin/apps/threads/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ThreadsConfig(AppConfig):
+    name = "threads"
