--- backend/channel_plugin/channel_plugin/users/urls.py
@@ -1,10 +1,0 @@
-from django.urls import include, path
-from rest_framework import routers
-from users.views import ThreadUserRolesViewset
-
-router = routers.DefaultRouter()
-router.register(r"", ThreadUserRolesViewset, basename="users")
-
-urlpatterns = [
-    path("thread_users_info/", include((router.urls, "users"))),
-]
