+++ backend/channel_plugin/config/urls.py
@@ -58,8 +58,10 @@ urlpatterns += [
     # API base url
     # path("api/", include("config.api_router")),
     path("api/", include("channel_plugin.info.urls")),
-    path("api/", include("channel_plugin.users.urls")),
     path("api/channels/", include("apps.channels.urls")),
+    path("api/channels/", include("apps.channelmessages.urls")),
+    path("api/channels/", include("apps.roles.urls")),
+    path("api/channels/", include("apps.threads.urls")),
     # DRF auth token
     # path("auth-token/", obtain_auth_token),
 ]
